Define core cmake_helper commands.
Added macro/function definitions to cmake_helper-config.cmake so that
new library and executable modules could be defined. The
cmh_end_module() macro was renamed to cmh_link_modules() to better
reflect its purpose and to avoid having the user think it was required
in header and library module CMakeLists.txt files.